Cleaned up code
git-svn-id: http://svn.code.sf.net/p/servestream/code/trunk@196 b8d320ad-ce4d-463a-8333-51e6c77ad906

+++ src/net/sourceforge/servestream/dbutils/StreamDatabase.java
@@ -116,13 +116,10 @@ public class StreamDatabase extends SQLiteOpenHelper {
 	
 	public void onRobustUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) throws SQLiteException {
 		
-		//if (oldVersion == 2) {
-			
+		if (oldVersion <= 2) {
 			db.execSQL("DROP TABLE IF EXISTS " + TABLE_STREAMS);
 			onCreate(db);
-			
-			//return;
-		//}
+		}
 	}
 	
 	/**
